backends/django-drf/pyproject.toml
[project]
name = "django-drf"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1.4",
    "djangorestframework-simplejwt>=5.3.1",
    "djangorestframework>=3.15.2",
    "django-filter>=24.3",
]

[dependency-groups]
dev = ["ruff>=0.8.3", "black>=24.10.0"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "benchboxd.test_settings"
